requirements.txt
et-xmlfile==1.0.1
future==0.18.2
jdcal==1.4.1
openpyxl==2.6.4
python-wialon==1.2.0
simplejson==3.17.2
xlrd==1.2.0


the human centipede,
hard candy,
a serbian film
the skin i live in
